rename reference implementation files
This commit renames generic/reference implementation
files to follow Go standard lib conventions.
